+++ examples/editor/src/PluginsConfig.js
@@ -75,19 +75,30 @@ export default {
     //Here you can call your custom video upload functionality (comment function to disable custom upload)
     handleFileSelection: (updateEntity, removeEntity) => {
       console.log('consumer wants to upload custom video');
-      const customVideoUrl = 'http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4';
+      const videoWithAbsoluteUrl = {
+        url: 'http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4'
+      }
+      const videoWithRelativeUrl = {
+        pathname: 'video/441c23_84f5c058e5e4479ab9e626cd5560a21b/file',
+        thumbnail: {
+          pathname: 'media/441c23_84f5c058e5e4479ab9e626cd5560a21bf000.jpg',
+          height: 1080,
+          width: 1920
+        }
+      };
+      // You can provide either absolute or relative URL.
+      // If relative URL is provided, a function 'getVideoUrl' will be invoked to form a full URL.
+      const videoToUpload = videoWithAbsoluteUrl;
       setTimeout(() => {
-        // You can provide either URL or pathname. If pathname is provided getVideoUrl will be invoked to form full URL.
-        updateEntity({ data: { url: customVideoUrl } });
+        updateEntity({ data: videoToUpload });
         //updateEntity({ error: { msg: 'Upload Failed' } });
-        console.log('consumer uploaded ', customVideoUrl);
+        console.log('consumer uploaded ', videoToUpload);
       }, 500);
     },
     enableCustomUploadOnMobile: true,
-    // Function is invoked when rendering video which only has pathname. You should take the pathname and form a full URL.
-    // getVideoUrl: src => {
-    //   return `https://video.wixstatic.com${src.pathname}`;
-    // },
+    // Function is invoked when rendering video which has relative URL.
+    // You should take the pathname and form a full URL.
+    getVideoUrl: src => `https://video.wixstatic.com/${src.pathname}`
   },
   [GIPHY_TYPE]: {
     giphySdkApiKey: "JfQziSTdnSwDZcD3w8DpUL4LMFu3zBgU"