+++ scripts/deployExamples.js
@@ -1,21 +1,7 @@
 /* eslint-disable no-console, fp/no-loops */
 
-const path = require('path');
 const chalk = require('chalk');
 const execSync = require('child_process').execSync;
-const sgAutorelease = require('surge-github-autorelease');
-
-const rootDir = path.resolve(__dirname, '..');
-
-const surgeOpts = {
-  repoOwner: 'wix-incubator',
-  repoName: 'rich-content',
-  rootPath: rootDir,
-};
-
-if (process.env.GITHUB_TOKEN) {
-  surgeOpts.githubToken = process.env.GITHUB_TOKEN;
-}
 
 const exec = cmd => execSync(cmd, { stdio: 'inherit' });
 
@@ -38,15 +24,15 @@ function build(example) {
   exec(buildCommand);
 }
 
-async function publish(example) {
+function publish(example) {
   console.log(chalk.cyan(`Publishing ${example.name} example v${example.version} to surge...`));
   const domain = `${example.name}-${example.version.replace(/\./g, '-')}`;
+  const deployCommand = `npx surge-github-autorelease -b . -s ${example.path}/dist -d ${domain}`;
   try {
-    await sgAutorelease({
-      ...surgeOpts,
-      domain,
-      sourceDirectory: `${example.path}/dist`,
-    });
+    console.log(
+      chalk.magenta(`Running "${deployCommand}`)
+    );
+    exec(deployCommand);
     console.log(chalk.green(`Published to ${fqdn(domain)}`));
   } catch (e) {
     console.error(chalk.bold.red(e));
@@ -66,9 +52,9 @@ async function deployExamples(examples) {
 
   for (const example of examples) {
     console.log(chalk.blue(`\nDeploying ${example.name} example...`));
-    bootstrap(example);
-    build(example);
-    await publish(example);
+    // bootstrap(example);
+    // build(example);
+    publish(example);
   }
 }